postgres_to_es/requirements.txt
elastic-transport==8.4.0
elasticsearch==8.5.3
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pydantic==1.10.2